Get bluetooth working
Posting encoder and button values over bluetooth now. Yay

Next up, add a status light for whether a client is connected.

Might also be nice to have the button count up instead of toggling. And maybe bump encoder to uint16. Not really necessary, it would be hard to spin it fast enough to have an ambiguous overflow or underflow I guess.