Refactor
Pulled volume stuff out into a separate module. Still a bit wonky but at least it's broken up a bit more. Could still use some improvement, if the server loses power than reconnects it will have reverted to it's initial value, which causes some confusion. We should cache the value and write it to the server on connect. That sounds like a lot though, and overall I'm very happy with it. It can lose and regain connection without stuttering at all. Awesome!