scratch
The one that records to a file works surprisingly well! The one that tries to do both didn't work, not sure why. Finally got the uncommented one working. Kept stuttering for the longest time, finally figured out it's because I left the log level at INFO. Duh. It says in the instructions not to do that. Switched it to warning and it works fine now. Doesn't bode well for adding WiFi though lol.

Anyway, committing messy just to capture some of the prototype stuff.