mqtt!
I got everything working pretty nicely. There's a slider in Home Assistant, it sets a volume percent int in the firmware, home assistant retains the value, and the device retrieves the value on reset. Woo!!! Ofc it's a huge mess, and I still need to figure out how to reconnect without blocking, and more importantly combine the mqtt with the audio and make sure they even can work at the same time, but we're getting there.