Ramping. Deployed.
Added 2 second ramp rate to all volume changes. No, ramp rate is not proportional to the amount the volume changed, it's always 2 seconds. Deal with it. It's good enough. It can be set via MQTT though I don't have that set up in home assistant and may never. Deployed to the actual noise maker in the nursery 2025-09-05

I had pulled the box because it kept turning on randomly when the mqtt server crashed and reset. After playing with it a bit, logging, and intentionally resetting it and the mqtt server, I'm pretty sure this is a problem with the mqtt server changing the value on a power cycle. I don't think the box itself is doing it. I think the mqtt server is coming up with a different volume status than what it had when it went down. I don't know why. Hopefully I can just get it to stop resetting...