Actual Fix
So my last fix didn't actually fix much. True, the MQTT server is changing the status when it resets, and I can't fix that. But I CAN keep servicing the audio player while waiting to reconnect instead of blocking, so that the audio doesn't drop out while the server is down then come back.

I also changed to only subscribe to status the first time the speaker boots up, rather than every time we reconnect. That way, if the MQTT server resets and comes back with the wrong value, we don't see it.

Also also, if the MQTT server does reset and come back up with the wrong value, once we reconnect, we'll write our value to it. But that's mutually exclusive: We only ask the server for status if we don't have any prior knowledge of the status, and we only reinforce the status to the server if we do have prior knowledge, and it may not.